0.4.0 release (#411)
* changes to docs

* Update cylon_release_0.4.0.md

* Update README.md

* changes to release note

* adding release notes to docs (#399)

* adding csv reader

* reorderd code snippets

* generated new frame.html

* adding docs changes

* adding group by

* updating docs

* changed front page code

* added missing "

* Adding group by (#405)

* reorderd code snippets

* generated new frame.html

* adding docs changes

* adding group by

* updating docs

* changed front page code

* added missing "

* partially fixing #375 (#406)

* partially fixing #375

* adding __len__ to df

* Bug fixes and cosmetic changes (#408)

* partially fixing #375

* adding __len__ to df

* passing pycylon.Table to df operations

* fixing buffer copy bug

* logging status errors in macros

* bug fix

* test case for chunked tables

* cosmetic changes

* fixed front page icons order

* Fixing 375 (#410)

* partially fixing #375

* adding __len__ to df

* passing pycylon.Table to df operations

* fixing buffer copy bug

* logging status errors in macros

* bug fix

* test case for chunked tables

* cosmetic changes

* fix #375

* fix #375

+++ cpp/src/cylon/join/join_utils.hpp
@@ -45,7 +45,7 @@ arrow::Status build_final_table_inplace_index(
     arrow::MemoryPool *memory_pool);
 
 arrow::Status CombineChunks(const std::shared_ptr<arrow::Table> &table,
-                            int64_t col_index,
+                            int col_index,
                             std::shared_ptr<arrow::Table> &output_table,
                             arrow::MemoryPool *memory_pool);
